split SLA from feature
